Extract `SwitchProductButton` component from `HeaderProduct`
