+++ src/components/index.ts
@@ -15,3 +15,4 @@ export * from "./LangSwitch";
 export * from "./HeaderAccount";
 export * from "./HeaderProduct";
 export * from "./AccountDropdown";
+export * from "./SwitchProductButton";
